before_lambda/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alns==7.0.0
asyncpg==0.31.0
attrs==25.4.0
bcrypt==5.0.0
branca==0.8.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
cykhash==2.0.1
Cython==3.2.3
dotenv==0.9.9
folium==0.20.0
fonttools==4.61.1
frozenlist==1.8.0
geopandas==1.1.1
idna==3.11
invoke==2.2.1
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
multidict==6.7.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
paramiko==2.11.0
pillow==12.0.0
propcache==0.4.1
psycopg==3.3.2
psycopg-binary==3.3.2
pyarrow==22.0.0
pycparser==2.23
PyMySQL==1.1.2
PyNaCl==1.6.1
pyogrio==0.12.1
pyparsing==3.2.5
pyproj==3.7.2
pyrobuf==0.9.3
pyrosm==0.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-rapidjson==1.23
pytz==2025.2
requests==2.32.5
scipy==1.16.3
shapely==2.1.2
six==1.17.0
sshtunnel==0.4.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
xmltodict==1.0.2
xyzservices==2025.11.0
yarl==1.22.0
